Safari-611.1.10/LayoutTests/platform/wk2/http/tests/media/video-auth-expected.txt

127.0.0.1:8000 - didReceiveAuthenticationChallenge - ProtectionSpaceAuthenticationSchemeHTTPBasic - Responding with username:password

Tests that the media player sends authorization credentials when requesting a media file.
Testing same domain (127.0.0.1)
EVENT(canplay)
END OF TEST
